+++ tests/core/simulation_control_03.cc
@@ -17,19 +17,20 @@ test()
 {
   Parameters::SimulationControl simulation_control_parameters;
 
-  simulation_control_parameters.dt     = 0.1;
-  simulation_control_parameters.adapt  = false;
-  simulation_control_parameters.maxCFL = 2;
+  simulation_control_parameters.dt             = 0.1;
+  simulation_control_parameters.adapt          = false;
+  simulation_control_parameters.adapt_with_cfl = false;
+  simulation_control_parameters.maxCFL         = 2;
   simulation_control_parameters.method =
     Parameters::SimulationControl::TimeSteppingMethod::bdf1;
   simulation_control_parameters.time_end = 0.5;
   simulation_control_parameters.output_control =
     Parameters::SimulationControl::OutputControl::iteration;
-  simulation_control_parameters.output_iteration_frequency             = 2;
-  simulation_control_parameters.time_step_independent_of_end_time      = true;
-  simulation_control_parameters.output_time_interval                   = {0,
-                                                                          1.7976931348623157e3};
-  simulation_control_parameters.respect_capillary_time_step_constraint = false;
+  simulation_control_parameters.output_iteration_frequency           = 2;
+  simulation_control_parameters.time_step_independent_of_end_time    = true;
+  simulation_control_parameters.output_time_interval                 = {0,
+                                                                        1.7976931348623157e3};
+  simulation_control_parameters.adapt_with_capillary_time_step_ratio = false;
 
   {
     SimulationControlTransient simulation_control(
@@ -157,6 +158,7 @@ test()
 
   {
     simulation_control_parameters.adapt                        = true;
+    simulation_control_parameters.adapt_with_cfl               = true;
     simulation_control_parameters.time_end                     = 20;
     simulation_control_parameters.dt                           = 1;
     simulation_control_parameters.adaptative_time_step_scaling = 1.2;
@@ -198,6 +200,7 @@ test()
   }
   {
     simulation_control_parameters.adapt                        = true;
+    simulation_control_parameters.adapt_with_cfl               = true;
     simulation_control_parameters.time_end                     = 25;
     simulation_control_parameters.dt                           = 1;
     simulation_control_parameters.adaptative_time_step_scaling = 1.2;
@@ -240,6 +243,7 @@ test()
 
   {
     simulation_control_parameters.adapt                        = true;
+    simulation_control_parameters.adapt_with_cfl               = true;
     simulation_control_parameters.time_end                     = 25;
     simulation_control_parameters.dt                           = 1;
     simulation_control_parameters.adaptative_time_step_scaling = 1.2;
@@ -282,6 +286,7 @@ test()
 
   {
     simulation_control_parameters.adapt                        = true;
+    simulation_control_parameters.adapt_with_cfl               = true;
     simulation_control_parameters.time_end                     = 25;
     simulation_control_parameters.dt                           = 1;
     simulation_control_parameters.adaptative_time_step_scaling = 1.2;